refactor: use rate limit buckets to determine slowmode
This allows a more accurate response to channel activity. This implementation is heavily inspired by, and is essentially 1:1 ported from, SlothBot by TPH (written by Princess Lana).

Thank you TPH for making your bot open source. The use of buckets fixed the issues I was having. You have been added to the LICENSE and the README since this was not my code.